Update patch set 1
Patch Set 1:

Should we have 2 extra headings

Availability of the Lab:
Lab providers can state the availability of the lab such as resources from the lab are connected to OPNFV CI and can be/is used for CI purposes or members of OPNFV community can login and use the lab for OPNFV purposes.

How to Get Access: Detailing the steps to take to get access.

But apart from these, it looks good.

Patch-set: 1