Updated Dell OPNFV Testlab info
Updated Dell OPNFV Tastlab info, including IP and access details to make it consistent. Also added figures in the right folder so that they could be embedded within the document. I was unable to delete the earlier images in doc/images. It would be great if these could be deleted by anyone having the necessary rights.

JIRA: PHAROS-14

Change-Id: I767d52913b09a6ba198a6325a1389f3660836a1f
Signed-off-by: Nauman Ahad <Nauman_Ahad@dell.com>