Update patch set 1
Patch Set 1:

the idea is to add them as committer (at least to create rst and template) not as contributor. I had a look at functest and releng, there is no contributor section in INFO file.
Contributor is defined in the wiki for people who will attend but not commit file in the repo, that is not the goal here

Patch-set: 1
Reviewer: Gerrit User 335 <335@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0