Created a pharos specification docment according to the new toolchain.
Created docs/specification and the index.rst fil in the new format
for the new toolchain sequence.  Will create a specification/index.html
link on artifacts and associate pdf document.
Made small editorials to the origincal content, but it needs work.

Left the original pharos-spec file as it being worked on still.
Should be removed once the spcification docs are in equivalent shape.

Change-Id: I6edb121766e7e1fdf1f38c70be95347b81b71dcc
Signed-off-by: ChristopherPrice <christopher.price@ericsson.com>