Update patch set 1
Patch Set 1:

Concerns about security are certainly valid however the intention of these templates is not to force any lab to publish information. It is up to each lab owner to find the balance between usability and security. I think this will be different for each lab depending on technical and non-technical factors. The templates are meant to help a lab efficiently and accurately document their configurations and this information can be used as a communication tool for e.g. when debugging issues with help of community or for projects to understand what the POD looks like (which has been a problem to date). We do expect labs to publish *some* useful information about what they are offering but again some specifics may be on a need to know basis which is fine IMO. Would it help to put something at the top of each template explaining the purpose and specifically cautioning NOT to publish information that could compromise security?

Patch-set: 1
Label: Verified=0