Update patch set 1
Patch Set 1: Code-Review-2

Hello.

Im sorry, i cannot +1,+2 this layout.

Over the past year+ we have worked to move POD information such as INFRA IP and Login information into a secured Lab repo so as to obfuscate the information in a certain way.

I am not sure about the Meeting on Sept 14th or a decision that was taken, since i dont see a quorum of PHAROS committers being present nor discussion about how this affects what is already in place and being used for the past two releases.

Patch-set: 1