Update patch set 2
Patch Set 2: Code-Review-1

What is the intention of this file? Is it to:
* broaden the naming of a scenario (in which case it might be ok), or
* serve as a configuration file

In case it is to serve as a config file, then we need to be more granular with regards to components and their configuration, i.e. you'd need to be able to configure things like:

* exact versions for each component (i.e. just "fdio" won't be sufficient - even version vs. master might not be sufficient).
* number of hugepages
* iommu support
* virt cpu pinning
* # of threads for galera db (if used)

etc.

BTW/ - please also consider what has been done in the past, i.e.
https://wiki.opnfv.org/display/genesis/Common+Deploy+Settings
https://wiki.opnfv.org/display/genesis/Common+Network+Settings

Patch-set: 2
Reviewer: Gerrit User 251 <251@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1