Update patch set 2
Patch Set 2:

(7 comments)

I have quite a few questions on the ambitions with this proposal. What is the purpose really, at the same time I see benefits providing some assumptions I spoke about on the Infra meeting earlier to day.
But lets make no mistake, unless this scenario definition file is exposed to/and in some way represents the underlying data model/schema for all the installers - the schema is likely to change with each new feature comming in, and so is the implementation of the necessary per installer translation engine below. So we are just pushing complexity one level down the stack.
See comments in-line, and apart from those in-line comments what I miss all-together below:
- Feature/package depency map - unless we want this to be hardcoded
  in the underlying translation engine. Which in case means that the
  scenario definition flexibility is just imaginary, in fact
  everything is bolted under the hood!
- In case the feature(s) within a scenario are services, we need to
  be able to assign them a role for co-location or separation
  what regards to resources.
- A generic model of configuring features - not part of the generic
  schema.
I hope you understand my comments, these are things I have had to tackle for the past two years, and is why Im careful in my judgements!

Patch-set: 2