LaaS development pod hosting using FOG
JIRA: N/A

This forms the base of the LaaS dev pod hosting. The pod_manager
selects a host from a pool of available machines and will image
the host, and the deployment manager will provision the host
and install OPNFV.

Adds the vpn manager in source/api.
The vpn file contains one formal interface definition
and one specific implementation, using LDAP.

Change-Id: I2b9abadea50874d8a9f9bd0d83fa06c7182b5cba
Signed-off-by: Parker Berberian <pberberian@iol.unh.edu>