LaaS development pod hosting using FOG
JIRA: N/A

This forms the base of the LaaS dev pod hosting. The pod_manager
selects a host from a pool of available machines and will image
the host, and the deployment manager will provision the host
and install OPNFV.

Adds the pharos dashboard api listener. This will periodically
poll the dashboard api and start a deployment is there is a
booking scheduled for one of your servers.
Also adds the listen.py script, which will start the pharos api
listener in the background, so it does not require a terminal
to run.

Change-Id: I2b9abadea50874d8a9f9bd0d83fa06c7182b5cba
Signed-off-by: Parker Berberian <pberberian@iol.unh.edu>