LaaS development pod hosting using FOG
JIRA: N/A

This forms the base of the LaaS dev pod hosting. The pod_manager
selects a host from a pool of available machines and will image
the host, and the deployment manager will provision the host
and install OPNFV.

Adds support for the Fuel installer. If requested, the
deployment_manager will launch the fuel installer, which will
install and configure Fuel onto the development pod, and
launch an instance of Openstack using the Fuel installation

Change-Id: I2b9abadea50874d8a9f9bd0d83fa06c7182b5cba
Signed-off-by: Parker Berberian <pberberian@iol.unh.edu>