LaaS development pod hosting using FOG
JIRA: N/A

This forms the base of the LaaS dev pod hosting. The pod_manager
selects a host from a pool of available machines and will image
the host, and the deployment manager will provision the host
and install OPNFV.

Adds the libvirt manager, which controls and creates
virtual machines and networks on the dev pod host.
Also adds the domain and network classes, which are simple
abstractions for a virtual machine and virtual network.

Change-Id: I2b9abadea50874d8a9f9bd0d83fa06c7182b5cba
Signed-off-by: Parker Berberian <pberberian@iol.unh.edu>