LaaS development pod hosting using FOG
JIRA: N/A

This forms the base of the LaaS dev pod hosting. The pod_manager
selects a host from a pool of available machines and will image
the host, and the deployment manager will provision the host
and install OPNFV.

Adds bash scripts in hostScripts/ which will be copied
over to the developemnt pod and used during the deployment
process.

Change-Id: I2b9abadea50874d8a9f9bd0d83fa06c7182b5cba
Signed-off-by: Parker Berberian <pberberian@iol.unh.edu>