Add Version Field to PDF
With Euphrates being the initial 'release' of the PDF it is important we
include a version specifier.

Without a version specifier, PDF consumers (computers or humans)
won't know which fields are be available, nor their possible type
(string, int, list, etc).

This also allows the PDF spec to be modified without fear of breaking
any tooling relying on the layout of fields or extance of specific
properties; Presuming tooling handles the version specifier.

Change-Id: I304ecc9ebfa0a36d40db8ef3486a3600ae86422e
Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>