Moving all pdfs to pharos repo
Inline encrytion of yaml files allows us to  make away with securedlab's
restrictions
Going forward, anyone will be able to see Lab PDFs

See:
https://github.com/opnfv/pharos/blob/master/config/utils/README.eyaml.rst

If you are the reviewer on this changeset it is becasue
git blamed showed you to be the author of the pdf in this review.

Encrypting ipmi password is optional. I have scrubbed the password from
this change. If you do not need the password to be hidden, please let me
know in the comment, and I will put it back.

If you do need the password hidden, please follow the readme above and
ammend this patch to include the encrypted value that you create on your
pods jumphost.

Change-Id: Ie3e935e7e9250a387858c29eb77c3c06b4fe3f57
Signed-off-by: agardner <agardner@linuxfoundation.org>