Moving all pdfs to pharos repo
Inline encrytion of yaml files allows us to make away with securedlab's
restrictions. Going forward, anyone will be able to see Lab PDFs

See:
https://github.com/opnfv/pharos/blob/master/config/utils/README.eyaml.rst

If you are the reviewer on this changeset it is becasue
git blamed showed you to be the author of the pdf in this review.

Encrypting ipmi password etc is optional. I have scrubbed the password
from this change. If you do not need the password to be hidden, please let me
know in the comment, and I will put it back.

If you do need the password hidden, please follow the readme above and
ammend this patch to include the encrypted value that you create on your
pods jumphost.

Change-Id: Idc70540477d8569f15335e75e7725c1e4e56a60b
Signed-off-by: agardner <agardner@linuxfoundation.org>