Update patch set 2
Patch Set 2:

@Parker Berberian
Interface and ports are not the same IMO. 
A port is the physical connector, what you see with your eyes on a server. An interface is the device the OS sees. Normally is not clear which port an OS interface is using and their order. This model tries to cover that by linking interface with server port. It also leverages on ports to group their vlan config and provide a known topology

Patch-set: 2
Label: Verified=0