validate-pdf-jinja2-templates-master failing
If keys do not exist, the skipping decryption python stuff doesn't work.
see:
https://build.opnfv.org/ci/job/validate-pdf-jinja2-templates-master/16/consoleFull

This change puts a dumb check for file with an if else statement.
Works on my machine.

Still does not fallback to plain in the case of a 'bad decryption'
but the current code does not eaither, bad decryption is logged
in any case.

Change-Id: I8da2c7cf8568a6e77ca24cc9ce665410d93b7ba9
Signed-off-by: agardner <agardner@linuxfoundation.org>