Auto Generated INFO.yaml file
A programic way of determining project lead is needed for the Testapi app
And will be used to grant ptl's additional ops

This file was auto generated from an ldap lookup and this projects INFO file

(part of the)spaghetti code:
https://gist.github.com/anonymous/85a395b22911ed9f04dd4bf61bff9dc2

This is as good as I can get the output
Project leads will need to be asked to clean up the rest

Let's discuss at the infra call before I spam all the repos.

INFO.yamls neeed to be committed in all repos before Serena Feng can
complete her work

Having a machine readble info file will also be useful in the future.

Change-Id: I562a486b70f4f0fecd0275c22fe4030097e39c3e
Signed-off-by: agardner <agardner@linuxfoundation.org>