[PDF] arm-pod*: Move net_config to IDF
While at it, drop 'fixed_ips', which is a leftover from early
net_config implementation and is not currently used by the Fuel
installer (only Daisy seems to use 'fixed_ips', but falls back
to using 'address' of the first interface when 'fixed_ips' is not
present).

Keep 'address: 192.168.11' on jumpserver first interface for Daisy
compatibility.

Change-Id: I83455a45a2a19fde6b9fb8b7f0fa8872b1291ab0
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>