Update patch set 1
Patch Set 1:

> (1 comment)

Yes, that was for IDF fuel bridge names, as I'm sure that handling those values accepts null values.

We can  go with ~ here too, instead of ''. We can update the schema to accept ['string', 'null'], but it puts some additional pressure on all installer teams to handle the null case gracefully. With '' we can just look for the feature we want using strstr.
Although, accepting null will also reduce some of the code churn in this topic.

Patch-set: 1