Update patch set 2
Patch Set 2:

> trick :)
 > 
 > Is this file not used in the real env?

In the real env, the jumpserver should have the eyaml keys and it should decrypt the PDF just fine.
However, on the build server where we run the verify job, the keys are not available and we always get the raw data, which leads to yamllint complaining about line too long (which is not a valid warning, as in the real env the descryped strings will be a lot shorter).
This is a trick, but it doesn't affect the real env in any way, only the verify job for pharos.

Patch-set: 2
Label: Verified=0