Update patch set 3
Patch Set 3:

> > Shall we add a top LICENSE file in the config/ sub-directory to
 > > avoid add the license info into all the files? This is also a
 > > simple way passing releng-anteater.
 > 
 > That's a good idea, sorry I only saw this comment after I rebased
 > it. I'll take a look at reworking this, maybe drop a LICENSE file
 > in each dir to allow different copyright for each installer dir.

On a closer look, it looks like the files that did not have license before this change should have different copyright headers (joid, apex, config/pdf are maintained by Canonical, Red Hat and Intel respectively).
So I'm not sure a global license header for the whole config/labs dirs is ok.
Anyway, I moved this change outside the `yamllint-fix` topic, so it can be discussed separately and doesn't block the merge of those changes. It's not critical and not a problem if we abandon this altogther.

Patch-set: 3
Reviewer: Gerrit User 2997 <2997@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0