Update patch set 2
Patch Set 2: Code-Review+2

> (1 comment)

Really good catch, Ting!
In order not to rebase ~20 patches for something that we don't yet use anyway (we don't do Daisy deploys on this pod), I suggest we merge this change as-is, then later add proper validation for IP addresses in the schema, and deal with this error in that patch.

Patch-set: 2
Reviewer: Gerrit User 2997 <2997@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+2