Update patch set 2
Patch Set 2:

> Hi?Alex?
 > Need I submit a patch to complete storage network or just make a
 > review to know what changes here?

Hi, Louie,
Until you actually deploy using Daisy on that POD, it is not mandatory to enable Daisy testing for it, although that is nice for completeness.
You can also add a dummy network (e.g. 10.10.10.10/24), just to keep the template check happy. That would have be to a real network when you actually deploy using Daisy.
So it's up to you if you want to add that section, but if you want to, we should handle it in a different change.

Patch-set: 2