Update patch set 6
Patch Set 6:

(1 comment)

Fatih, i know you point of view, XCI is not an installer, but we don't have to be as much dogmatic in the code.

idf.xci is a key in a dict, containing all data needed by the tools/scripts/code used by the xci project to deploy the target scenarios. As an example idf.xci.pxe_network is used by bifrost not by any installer.

If we use OSA, What will come next, we will duplicate all the config for Kolla ? for k8s ? for FooBar ? 

I am opposed to idf.osa and think idf.xci.installer:{} is a good alternative to contains what is specific to each installer and idf.xci can contains all that is common to all of them.

Patch-set: 6
Label: Verified=0