[fuel] IDF: Add DPDK parameters support
In order to enable DPDK for OVS scenarios in Armband, we need DPDK
specific configuration to be parameterizable.
The default DPDK configuration values will remain in Fuel repo and
will be overriden with the values defined in IDF (only if present).
Since hugepage config and CPU pinning also apply to non-DPDK
scenarios, handle them together for 'common', respectively 'dpdk'.

To keep things simple, we will reuse the format expected by Fuel's
reclass model, so we can pass the whole config block as-is.

- IDF schema: fuel: Add strict checking of new block's structure,
  while allowing the rvalues to be more or less freestyle;
- arm-pod6 IDF: Add initial config block implementation;

JIRA: ARMBAND-342

Change-Id: I9224f06bb6b9fb2387414eb2a1be35461c27fb99
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>