Update patch set 20
Patch Set 20: Code-Review+1

I haven't checked this hands-on (i.e. I haven't generated a pod_config.yml to see how it looks), so I trust you this is tested and will work :)
Imo, it's still not perfect, but it's a lot better than what we had before.
Before we merge it, can we see some example outputs? e.g. for PDFs with /24 public and /25 public
If nobody objects, we will merge this soon.

Patch-set: 20
Label: Code-Review=+1