Update patch set 2
Patch Set 2:

Sorry, I was a bit busy with the release and didn't get a chance to look at this yet.
The first thing I noticed is that this change only adds the jumpserver, although I was expected some 5 dummy nodes (pdf nodes have the type param, which can be 'baremetal' or 'virtual') - see [1, 2] for what I had in mind.
I'm ok with merging this as-is, but I'd like to discuss adding those dummy nodes, it'd make things easier for installers to control the number (and later roles too) of virtual nodes via PDF/IDF/SDF.
[1] https://gerrit.opnfv.org/gerrit/#/c/50315/
[2] https://gerrit.opnfv.org/gerrit/#/c/51265/

Patch-set: 2