Update patch set 2
Patch Set 2:

> > anteater is not smart enough and that's a false positive.
 > > We use safe_load by loading the C-version or the python version
 > of
 > > the safe loader via:
 > > LOADER = yaml.CSafeLoader if yaml.__with_libyaml__ else
 > > yaml.SafeLoader
 > 
 > Hi Alex, do you suggest to use a new way instead of the
 > yaml.safe_load? or we just keep it?

Imo, anteater should be smarter and figure out we override the default loader to the safe version in a different place, BUT that would be tricky to implement.
I would go for an anteater exception for these two occurences, because afaict they are false positives.
So I'd leave these as they are right now in pharos.

Patch-set: 2