Update patch set 1
Patch Set 1:

(1 comment)

Hi,
This looks mostly good, but there a couple of things to fix before merging:
- some minor yamllint violations (I'll push a new revision taking care of those, since they're trivial);
- license year and comany (ditto, I'll fix that in a new revision);
- possibly wrong interface names / busaddr in IDF (see inline comment for that, you will need to check the nodes and see which one is wrong);
- filenames should respect the Pharos convention (i.e. PDF should be named podX.yaml, while IDF should be idf-podX.yaml; I will fix this in revision 2);
One thing to keep in mind though - if you use Fuel as an installer on this POD and assign the 10.20.0.0/24 network to PXE/admin, you will need to set a different INSTALLER_IP when deploying.
That's because Fuel creates a very isolated network (only used on the jumpserver, so it doesn't need any extra configuration on the jumpserver or the switch/router) that it will use to connect to the installer node. By default INSTALLER_IP is 10.20.0.2, so the installer creates the mcpcontrol virsh network as 10.20.0.0/24.
To avoid such collisions, simply use a different INSTALLER_IP (e.g. 10.20.1.0/24) or switch PXE/admin in PDF/IDF to a different CIDR.

Patch-set: 1