Add port to remote management parameters
This change proposes addition of a new optional parameter to
PDF schema to add port to remote management parameters so one
can use none-default ports if needed.

Apart from the PODs that use none-default ports for remote
managamene, n example to this use case is the virtual deployments
that use virtualbmc for simulating BMC which requires calculation
of ports on the fly which could be messy. By adding the port into
PDF, it is straightforward to configure the tools to use none-default
ports if needed.

Change-Id: I4ce41c3977470576296a13dc668dec246fc29314
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>